Added primitive drawing functions: TESTED

+++ GameEngine/Display.hpp
@@ -9,7 +9,7 @@
 #include <SDL3/SDL.h>
 #include <iostream>
 
-#define POSITION(x,y) uint8_t x, uint8_t y
+#define POSITION(x,y) unsigned int x, unsigned int y
 #define COLOR uint8_t r, uint8_t g, uint8_t b
 
 constexpr int SCREEN_WIDTH = 320;
